Squash-commit of datadumper branch.

+++ run_chewiesnake.py
@@ -4,6 +4,8 @@ import pathlib
 import shutil
 import subprocess
 
+from data_handling import get_allele_profiles
+
 CHEWIESNAKE_IMAGE_ID = os.getenv("CHEWIESNAKE_IMAGE_ID")
 CHEWIESNAKE_MOUNT_POINT = os.getenv("CHEWIESNAKE_MOUNT_POINT")
 CHEWIESNAKE_OUTPUT_SUBFOLDER = os.getenv("CHEWIESNAKE_OUTPUT_SUBFOLDER", "output")
@@ -12,6 +14,7 @@ BIFROST_DB_KEY = os.getenv("BIFROST_DB_KEY", "mongodb://localhost/bifrost_test")
 
 parser = argparse.ArgumentParser(description='Run chewieSnake with selected samples and save allele profiles to MongoDB.')
 parser.add_argument('-s','--sample_names', nargs='+', help='Sample names (strip read numbers and extension from file names).')
+parser.add_argument('-d','--delete_old_output', action='store_true', help='Delete old output folder')
 args = parser.parse_args()
 
 lines = ["sample\tfq1\tfq2\n"]
@@ -25,7 +28,7 @@ for sample_name in args.sample_names:
 mount_point = pathlib.Path(CHEWIESNAKE_MOUNT_POINT)
 assert mount_point.exists()
 output_subfolder = pathlib.Path(mount_point, CHEWIESNAKE_OUTPUT_SUBFOLDER)
-if output_subfolder.exists():
+if args.delete_old_output and output_subfolder.exists():
     print("Output subfolder already exists and will be deleted!")
     shutil.rmtree(output_subfolder)
 samples_tsv_path = pathlib.Path(mount_point, 'samples.tsv')
@@ -58,9 +61,6 @@ process_out, process_error = process.communicate()
 """Note: if there's only one sample you'll always get an error from 
 Grapetree, but it doesn't matter in our case."""
 
-# Read relevant results from output subfolder
-assert output_subfolder.exists()
-allele_profiles_file = pathlib.Path(output_subfolder, 'cgmlst', 'allele_profiles.tsv')
-assert allele_profiles_file.exists()
-hashids_file = pathlib.Path(output_subfolder, 'cgmlst', 'hashids.tsv')
-assert hashids_file.exists()
+
+# Update allele_profiles collection in MongoDB
+get_allele_profiles(output_subfolder)
